requirements.txt
pandas
requests
beautifulsoup4
brotlipy
pyyaml
